refactor(ocale): migrate developer-products wire code into domains/developer-products
Move the wire code for the three Operations on
`/developer-products/v2/...` (`get`, `create`, `update`) from the peer
Resource folder `resources/developer-products/` into the by-Domain
layout under `domains/developer-products/products/` per ADR-022. The
Resource folder collapses to its public surface (`client.ts`,
`index.ts`, `index.spec-d.ts`); builders, parsers, operation-limit
constants, wire types, and parsed shapes now live in the Domain
sub-tree. The local `toBlob` helper in `builders.ts` is dropped in
favour of the shared `internal/utils/to-blob.ts` already used by the
existing `game-internationalization` Domain.

Wire requests, parsed shapes, and rate-limit operation keys
(`developer-products.{get,create,update}`) are preserved end-to-end.
Conformance tests, integration tests, and the type-test file are
updated to import from the new locations without behavioural change.
