+++ packages/open-cloud/src/domains/developer-products/products/parsers.spec.ts
@@ -1,7 +1,7 @@
 import { validDeveloperProductBody } from "#tests/helpers/developer-products";
 import { assert, describe, expect, it } from "vitest";
 
-import { ApiError } from "../../errors/api-error.ts";
+import { ApiError } from "../../../errors/api-error.ts";
 import { parseDeveloperProductResponse } from "./parsers.ts";
 import type { DeveloperProductConfigV2 } from "./wire.ts";
